+++ lib/src/answer_format/image_answer_format.g.dart
@@ -8,12 +8,18 @@ part of 'image_answer_format.dart';
 
 ImageAnswerFormat _$ImageAnswerFormatFromJson(Map<String, dynamic> json) =>
     ImageAnswerFormat(
-      defaultValue: json['defaulValue'] as String?,
-      buttonText: json['hint'] as String? ?? '',
+      defaultValue: json['defaultValue'] as String?,
+      buttonText: json['buttonText'] as String? ?? 'Image: ',
+      useGallery: json['useGallery'] as bool? ?? true,
+      hintImage: json['hintImage'] as String?,
+      hintTitle: json['hintTitle'] as String?,
     );
 
 Map<String, dynamic> _$ImageAnswerFormatToJson(ImageAnswerFormat instance) =>
     <String, dynamic>{
       'defaultValue': instance.defaultValue,
-      'hint': instance.buttonText,
+      'buttonText': instance.buttonText,
+      'useGallery': instance.useGallery,
+      'hintImage': instance.hintImage,
+      'hintTitle': instance.hintTitle,
     };