+++ lib/src/views/double_answer_view.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:survey_kit/src/answer_format/double_answer_format.dart';
-import 'package:survey_kit/src/views/decoration/input_decoration.dart';
 import 'package:survey_kit/src/result/question/double_question_result.dart';
 import 'package:survey_kit/src/steps/predefined_steps/question_step.dart';
+import 'package:survey_kit/src/views/decoration/input_decoration.dart';
 import 'package:survey_kit/src/views/widget/step_view.dart';
 
 class DoubleAnswerView extends StatefulWidget {
@@ -66,7 +66,7 @@ class _DoubleAnswerViewState extends State<DoubleAnswerView> {
       title: widget.questionStep.title.isNotEmpty
           ? Text(
               widget.questionStep.title,
-              style: Theme.of(context).textTheme.headline2,
+              style: Theme.of(context).textTheme.displayMedium,
               textAlign: TextAlign.center,
             )
           : widget.questionStep.content,
@@ -75,6 +75,7 @@ class _DoubleAnswerViewState extends State<DoubleAnswerView> {
         child: Container(
           width: MediaQuery.of(context).size.width,
           child: TextField(
+            autofocus: true,
             decoration: textFieldInputDecoration(
               hint: _doubleAnswerFormat.hint,
             ),