+++ lib/src/views/multiple_double_answer_view.dart
@@ -82,7 +82,7 @@ class _MultipleDoubleAnswerViewState extends State<MultipleDoubleAnswerView> {
       title: widget.questionStep.title.isNotEmpty
           ? Text(
               widget.questionStep.title,
-              style: Theme.of(context).textTheme.headline2,
+              style: Theme.of(context).textTheme.displayMedium,
               textAlign: TextAlign.center,
             )
           : widget.questionStep.content,
@@ -94,7 +94,7 @@ class _MultipleDoubleAnswerViewState extends State<MultipleDoubleAnswerView> {
               padding: const EdgeInsets.only(bottom: 32.0),
               child: Text(
                 widget.questionStep.text,
-                style: Theme.of(context).textTheme.bodyText2,
+                style: Theme.of(context).textTheme.bodyMedium,
                 textAlign: TextAlign.center,
               ),
             ),
@@ -108,6 +108,8 @@ class _MultipleDoubleAnswerViewState extends State<MultipleDoubleAnswerView> {
                     .entries
                     .map((MapEntry<int, String> md) {
                   return TextField(
+                    textInputAction: TextInputAction.next,
+                    autofocus: true,
                     decoration: InputDecoration(
                       labelText: md.value,
                     ),